% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wea_ames_30yrs}
\alias{wea_ames_30yrs}
\title{Historical weather data (1990-2019) in Ames, Iowa}
\format{A data frame with 10,957 rows and 10 variables:
\describe{
\item{station}{a unique 6 digit station code, alpha-numeric}
\item{station_name}{a unique name assigned to each station code that may be more informative, alpha-numeric}
\item{lat}{latitude, N}
\item{lon}{longitude, W}
\item{day}{date of measurement, ymd}
\item{doy}{day of year, numeric}
\item{high}{highest recorded temperature on that date, F}
\item{low}{lowest recorded temperature on that date, F}
\item{precip}{total precipitation for that date (rain equivalent), inches}
\item{merra_srad}{total radiation for that date, MJ per m2}
...
}}
\source{
\url{https://mesonet.agron.iastate.edu/request/coop/fe.phtml}
}
\usage{
wea_ames_30yrs
}
\description{
A dataset with weather data for Ames Iowa as downloaded from the Iowa Mesonet NWSCoop network.
}
\keyword{datasets}
